Return false so not to add extra entry to history
If you have a link whose href is "#" and you set up an event handler
which, on click, calls router.navigate(..., trigger: true), you will end
up with two entries in the history instead of one (one for the #,
another for the URL change). Returning false from the event handler
fixes this.